+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 1.1.0 Oct 30, 2020
 
 ### New
+- ChaCha20 encryption support `crypto.chacha20`. 
 - `boc.parse_shardstate` function for shardstates parsing
 
 ## 1.0.0 Oct 27, 2020